+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 * Report error when `install.sh` is run not from `build` folder
 * Add missing columns on write (upgrade schema) in DbFile
 * Add _iterator to DbFile id_template config
-* Fix space in DbFile cols in Schema
+* Fix space in DbFile cols/cols_optional in Schema
+* Fix typings in DbFile, Loggr, app_utils, os_utils
 
 ## 0.0.21 (2024-03-26)